requirements.txt
fastapi
gunicorn
uvicorn
openai
gitpython